lesson2/pyproject.toml
[project]
name = "lesson2"
version = "0.1.0"
description = "重作的專案"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
]
